requirements.txt
gradio==5.29.1
ffmpeg-python==0.2.0
openai-whisper==20240930
# torch==2.4.1 --index-url https://download.pytorch.org/whl/cu124
